requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==4.2
django-coverage-plugin==1.3.1
Sphinx==1.5
django-extensions==1.7.5
Werkzeug==0.11.11
django-test-plus==1.0.16
factory_boy==2.7.0

# Temporary workaround to fix issue https://github.com/pydanny/cookiecutter-django/issues/827
# It will be removed when django-debug-toolbar publish the new version on pypi
git+https://github.com/jazzband/django-debug-toolbar.git@637b0b73bc684f6ea600dca5ca1bb1f090fcee93#egg=django-debug-toolbar
#django-debug-toolbar==1.5

# improved REPL
ipdb==0.10.1

pytest-django==3.1.2
pytest-sugar==0.8.0
